Moved Copy() in env source file
